refactor: apply config and refactor hovering toolbar
