+++ packages/plugins/text/src/types/hovering-toolbar.ts
@@ -1,5 +1,6 @@
 import { ListsEditor } from '@prezly/slate-lists'
-import { BaseEditor } from 'slate'
+import { MouseEvent } from 'react'
+import { Editor as SlateEditor, BaseEditor } from 'slate'
 import { ReactEditor } from 'slate-react'
 
 import { TextPluginConfig } from './config'
@@ -8,6 +9,26 @@ export interface HoveringToolbarProps {
   closeSubMenuIcon: React.ReactNode
   closeSubMenuTitle: string
   config: TextPluginConfig
+}
+
+export interface HoveringToolbarControlsProps {
+  editor: SlateEditor
+  config: TextPluginConfig
+  closeSubMenuIcon: React.ReactNode
+  closeSubMenuTitle: string
+}
+
+export interface HoveringToolbarButtonProps {
+  editor: SlateEditor
+  config: TextPluginConfig
+  index?: number
+  control: TextEditorControl
+  setSubMenu: React.Dispatch<React.SetStateAction<number | undefined>>
+  onMouseDown(
+    event: MouseEvent,
+    control: TextEditorControl,
+    index?: number
+  ): void
 }
 
 export type TextEditorControl = ControlButton | NestedControlButton
